Config tree store data updated when options are saved.
